requirements.txt
python-telegram-bot==20.7
ollama>=0.1.0
chromadb==0.5.3
requests==2.31.0
beautifulsoup4==4.12.3
python-dotenv==1.0.0
psutil>=5.9.0
httpx>=0.27.0
aiohttp>=3.9.0